+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.21.9
+- Fixed Dragon Balls and Chaos Emeralds not being destroyed on extraction
+- Fixed Hellgato not changing value when non-hosts use it
+
 ## 0.21.8
 - Slightly changed Smith Note logging
 - Potential fix for Hellgato impact effects not playing